Fix motion and its editor gui

+++ Source/common/resource/anim_fsm/anim_fsm_transition.hpp
@@ -17,7 +17,7 @@ struct AnimFSMTransition {
         NOT_EQUAL
     };
     struct Condition {
-        size_t param_hdl;
+        int param_hdl;
         std::string param_name;
         CONDITION type;
         float ref_value;
